backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.3
motor==3.3.2
python-dotenv==1.0.1
httpx==0.26.0
